# openapi.yaml
openapi: 3.1.0
info:
  title: Jitsi Portal API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/rooms/{roomId}:
    get:
      summary: Get room by ID
      operationId: getRoom
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Room details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '401':
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '403':
          description: Access denied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '404':
          description: Room not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
    put:
      summary: Update room
      operationId: updateRoom
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomRequest'
      responses:
        '200':
          description: Room updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '400':
          description: Invalid request or config set
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '401':
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '403':
          description: Access denied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '404':
          description: Room not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '409':
          description: Room name conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
    delete:
      summary: Delete room
      operationId: deleteRoom
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Room deleted successfully
        '401':
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '403':
          description: Access denied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '404':
          description: Room not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '409':
          description: Room has active meetings
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
...
  /api/v1/meetings/{meetingId}:
    get:
      summary: Get meeting by ID
      operationId: getMeeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Meeting details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingResponse'
        '401':
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '403':
          description: Access denied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '404':
          description: Meeting not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
    put:
      summary: Update meeting
      operationId: updateMeeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeetingRequest'
      responses:
        '200':
          description: Meeting updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingResponse'
        '400':
          description: Invalid request or schedule
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '401':
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '403':
          description: Access denied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '404':
          description: Meeting not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '409':
          description: Meeting finalized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
...
  /api/v1/meetings/{meetingId}/invites:
    post:
      summary: Create invite for a meeting
      operationId: createMeetingInvite
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteRequest'
      responses:
        '201':
          description: Invite created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteResponse'
        '400':
          description: Invalid request payload
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '401':
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '403':
          description: Access denied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '404':
          description: Meeting not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '409':
          description: Invite conflict or already revoked
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '410':
          description: Invite expired or revoked
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
    get:
      summary: List invites by meeting
      operationId: listMeetingInvites
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paged invites list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedInviteResponse'
        '401':
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '403':
          description: Access denied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '404':
          description: Meeting not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
  /api/v1/meetings/{meetingId}/invites/bulk:
    post:
      summary: Create meeting invites in bulk
      operationId: createMeetingInvitesBulk
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkCreateInviteRequest'
      responses:
        '201':
          description: Bulk invite operation completed without validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkInviteResponse'
        '400':
          description: Invalid request payload
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '401':
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '403':
          description: Access denied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '404':
          description: Meeting not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '409':
          description: Meeting state does not allow issuing invites
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '422':
          description: Partial validation failure (some rows are invalid)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
  /api/v1/meetings/{meetingId}/invites/{inviteId}:
    delete:
      summary: Revoke invite by ID
      operationId: revokeMeetingInvite
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
        - name: inviteId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Invite revoked successfully
        '401':
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '403':
          description: Access denied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '404':
          description: Meeting or invite not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '409':
          description: Invite already revoked
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
...
components:
  schemas:
...
    ProblemDetailResponse:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        properties:
          type: object
          properties:
            errorCode:
              type: string
            traceId:
              type: string
          required:
            - errorCode
            - traceId
...
    UpdateRoomRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        configSetId:
          type: string
    RoomResponse:
      type: object
      properties:
        roomId:
          type: string
        name:
          type: string
        description:
          type: string
        tenantId:
          type: string
        configSetId:
          type: string
        status:
          type: string
          enum: [active, closed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - roomId
        - name
        - tenantId
        - configSetId
        - status
        - createdAt
        - updatedAt
...
    UpdateMeetingRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        meetingType:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        allowGuests:
          type: boolean
        recordingEnabled:
          type: boolean
    MeetingResponse:
      type: object
      properties:
        meetingId:
          type: string
        roomId:
          type: string
        title:
          type: string
        description:
          type: string
        meetingType:
          type: string
        configSetId:
          type: string
        status:
          type: string
          enum: [scheduled, canceled, ended]
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        allowGuests:
          type: boolean
        recordingEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - meetingId
        - roomId
        - title
        - meetingType
        - configSetId
        - status
        - startsAt
        - endsAt
        - allowGuests
        - recordingEnabled
        - createdAt
        - updatedAt
...
    CreateInviteRequest:
      type: object
      properties:
        role:
          type: string
          enum: [participant, moderator]
        maxUses:
          type: integer
          minimum: 1
        expiresInHours:
          type: integer
          minimum: 1
          maximum: 168
      required:
        - role
    BulkInviteRecipient:
      type: object
      properties:
        email:
          type: string
        userId:
          type: string
        role:
          type: string
          enum: [participant, moderator]
        rowIndex:
          type: integer
      description: Exactly one of email or userId should be provided.
    BulkCreateInviteRequest:
      type: object
      properties:
        recipients:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/BulkInviteRecipient'
        defaultTtlMinutes:
          type: integer
          minimum: 1
          maximum: 10080
        defaultMaxUses:
          type: integer
          minimum: 1
        defaultRole:
          type: string
          enum: [participant, moderator]
        duplicatePolicy:
          type: string
          enum: [SKIP_EXISTING, REPLACE]
      required:
        - recipients
    BulkInviteCreatedItem:
      type: object
      properties:
        inviteId:
          type: string
        token:
          type: string
        recipient:
          type: string
        role:
          type: string
          enum: [participant, moderator]
      required:
        - inviteId
        - token
        - recipient
        - role
    BulkInviteSkippedItem:
      type: object
      properties:
        recipient:
          type: string
        reason:
          type: string
      required:
        - recipient
        - reason
    BulkInviteErrorItem:
      type: object
      properties:
        rowIndex:
          type: integer
        recipient:
          type: string
        errorCode:
          type: string
        message:
          type: string
      required:
        - rowIndex
        - errorCode
        - message
    BulkInviteSummary:
      type: object
      properties:
        total:
          type: integer
        created:
          type: integer
        skipped:
          type: integer
        failed:
          type: integer
      required:
        - total
        - created
        - skipped
        - failed
    BulkInviteResponse:
      type: object
      properties:
        created:
          type: array
          items:
            $ref: '#/components/schemas/BulkInviteCreatedItem'
        skipped:
          type: array
          items:
            $ref: '#/components/schemas/BulkInviteSkippedItem'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/BulkInviteErrorItem'
        summary:
          $ref: '#/components/schemas/BulkInviteSummary'
      required:
        - created
        - skipped
        - errors
        - summary
    InviteResponse:
      type: object
      properties:
        id:
          type: string
        meetingId:
          type: string
        token:
          type: string
        role:
          type: string
          enum: [participant, moderator]
        maxUses:
          type: integer
        usedCount:
          type: integer
        expiresAt:
          type: string
          format: date-time
          nullable: true
        revokedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        valid:
          type: boolean
      required:
        - id
        - meetingId
        - token
        - role
        - maxUses
        - usedCount
        - createdAt
        - createdBy
        - valid
    PagedInviteResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/InviteResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
      required:
        - items
        - page
        - size
        - totalElements
        - totalPages
...